LPS-52484 Rename (no logical changes)
