+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-values-query/build.xml
@@ -1,0 +1,38 @@
+<?xml version="1.0"?>
+<!DOCTYPE project>
+
+<project>
+	<import file="../../../../tools/sdk/build-common-osgi-plugin.xml" />
+
+	<property name="auto.deploy.dir" value="${liferay.home}/osgi/modules" />
+
+	<target name="build-ddm-form-values-query-parser">
+		<delete>
+			<fileset
+				dir="src/com/liferay/dynamic/data/mapping/form/values/query/impl/parser"
+				excludes="*.ignore"
+				includes="*.*"
+			/>
+		</delete>
+
+		<java
+			classname="org.antlr.v4.Tool"
+			classpathref="plugin.classpath"
+			dir="src/com/liferay/dynamic/data/mapping/form/values/query/impl/grammar"
+			fork="true"
+			maxmemory="1024m"
+			newenvironment="true"
+		>
+			<arg value="-o" />
+			<arg value="../parser" />
+			<arg value="DDMFormValuesQuery.g" />
+		</java>
+
+		<fixcrlf
+			includes="*.java"
+			javafiles="yes"
+			srcdir="src/com/liferay/dynamic/data/mapping/form/values/query/impl/parser"
+			tab="add"
+		/>
+	</target>
+</project>
