--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-query/bnd.bnd
@@ -1,13 +1,0 @@
-Bundle-Name: Liferay Dynamic Data Mapping Query
-Bundle-SymbolicName: com.liferay.dynamic.data.mapping.query
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.dynamic.data.mapping.query.api
-Import-Package:\
-	!javax.imageio.*,\
-	!javax.print.*,\
-	!javax.swing.*,\
-	*
-Include-Resource:\
-	classes,\
-	@lib/antlr4-runtime.jar,\
-	@lib/org.abego.treelayout.core.jar
