--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-query/src/com/liferay/dynamic/data/mapping/query/impl/model/DDMFormFieldValueMatchesAnyMatcher.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.dynamic.data.mapping.query.impl.model;
-
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * @author Pablo Carvalho
- */
-public class DDMFormFieldValueMatchesAnyMatcher
-	implements DDMFormFieldValueMatcher {
-
-	public void addDDMFormFieldValueMatcher(
-		DDMFormFieldValueMatcher ddmFormFieldValueMatcher) {
-
-		_ddmFormFieldValueMatchers.add(ddmFormFieldValueMatcher);
-	}
-
-	@Override
-	public boolean matches(DDMFormFieldValue ddmFormFieldValue) {
-		for (DDMFormFieldValueMatcher ddmFormFieldValueMatcher :
-				_ddmFormFieldValueMatchers) {
-
-			if (ddmFormFieldValueMatcher.matches(ddmFormFieldValue)) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	private final List<DDMFormFieldValueMatcher> _ddmFormFieldValueMatchers =
-		new ArrayList<>();
-
-}
